Add blotto_kernel_lib.
Add basic blotto game primities such as Game, Army, User_token etc.

Users will authenticate with the one-time genrated token. This token
is unique and can be retrieved only once for a given (name, surname,
email) triplet. This is of course very vurnerable, but should be
sufficient for our usecases.